requirements.txt
Flask
Flask-RESTful
Flask-Cors
websocket-client